bot/requirements.txt
python-binance
python-telegram-bot
schedule
python-dotenv